Display actual daily expense amount on each calendar day.

+++ api/models/models.go
@@ -22,13 +22,13 @@ type UserAccount struct {
 }
 
 type UserExpense struct {
-	ID                 int64  `json:"id" db:"id"`
-	UserAccountID      int64  `json:"user_account_id,omitempty" db:"user_account_id"`
-	ExpenseType        string `json:"expense_type,omitempty" db:"expense_type"`
-	ExpenseCategory    string `json:"expense_category,omitempty" db:"expense_category"`
-	ExpenseDescription string `json:"expense_description,omitempty" db:"expense_description"`
-	ExpenseAmount      int64  `json:"expense_amount,omitempty" db:"expense_amount"`
-	ExpenseDate        string `json:"expense_date,omitempty" db:"expense_date"`
-	CreatedAt          string `json:"created_at,omitempty" db:"created_at"`
-	UpdatedAt          string `json:"updated_at,omitempty" db:"updated_at"`
+	ID                 int64   `json:"id" db:"id"`
+	UserAccountID      int64   `json:"user_account_id,omitempty" db:"user_account_id"`
+	ExpenseType        string  `json:"expense_type,omitempty" db:"expense_type"`
+	ExpenseCategory    string  `json:"expense_category,omitempty" db:"expense_category"`
+	ExpenseDescription string  `json:"expense_description,omitempty" db:"expense_description"`
+	ExpenseAmount      float64 `json:"expense_amount,omitempty" db:"expense_amount"`
+	ExpenseDate        string  `json:"expense_date,omitempty" db:"expense_date"`
+	CreatedAt          string  `json:"created_at,omitempty" db:"created_at"`
+	UpdatedAt          string  `json:"updated_at,omitempty" db:"updated_at"`
 }
